+++ CompositePruebaDulces/Contracts/IGenericBuilderProducto.cs
@@ -4,11 +4,11 @@ namespace Domain.Contracts
 {
     public interface IGenericBuilderProducto<out T> where T : Producto
     {
-		public T SetCostoUnitario(double costo);
-		public T SetCantidad(double cantidad);
+		public T SetCostoUnitario(decimal costo);
+		public T SetCantidad(decimal cantidad);
 		public T SetNombre(string nombre);
 		public T SetUnidadDeMedida(UnidadDeMedida unidadDeMedida);
-		public T SetPorcentajeDeUtilidad(double porcentajeDeUtilidad);
+		public T SetPorcentajeDeUtilidad(decimal porcentajeDeUtilidad);
 		public T SetEspecificacion(Especificacion especificacion);		
 		public T SetTipo(Tipo tipo);
 	}
